refactor: move tycho decoders into the individual protocol modules
