+++ src/protocol/mod.rs
@@ -7,7 +7,6 @@ pub mod errors;
 pub mod events;
 pub mod models;
 pub mod state;
-pub mod tycho;
 pub mod uniswap_v2;
 pub mod uniswap_v3;
 pub mod vm;
