+++ src/protocol/uniswap_v2/mod.rs
@@ -2,3 +2,4 @@
 pub mod events;
 pub mod reserve_price;
 pub mod state;
+pub mod tycho_decoder;
