+++ src/protocol/uniswap_v3/mod.rs
@@ -8,3 +8,4 @@ pub mod state;
 mod swap_math;
 pub mod tick_list;
 mod tick_math;
+pub mod tycho_decoder;
